Prepare to convert SBThreatType to an enum class in /components/subresource_filter
It's somewhat difficult to log an SBThreatType as an UMA histogram since
we don't have a kMaxValue. We can't simply add a kMaxValue because
SBThreatType is an enum instead of enum class, and we'd be exposing a
constant safe_browsing::kMaxValue. This CL prepares to convert
SBThreatType to an enum class in /components/subresource_filter.

That requires two changes:

- SBThreatType is no longer implicitly convertible to int, so we will
  need a few static casts

- SBThreatType values are now in the namespace
  safe_browsing::SBThreatType instead of just safe_browsing. In some
  places this is fixed by adding the new namespace, and in some places
  with a "using enum" declaration. We can't just add a "using enum" for
  all usages of SBThreatType because "using enum" would declare a
  constant kMaxValue. So "using enum" is only used when many
  SBThreatType constants are used, and in as narrow a scope as possible.

This CL was uploaded by git cl split.

Bug: 329698341
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5375397
Commit-Queue: Alex Turner <alexmt@chromium.org>
Auto-Submit: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1274145}

+++ components/subresource_filter/content/browser/subresource_filter_safe_browsing_client.cc
@@ -23,7 +23,7 @@ std::unique_ptr<base::trace_event::TracedValue>
 SubresourceFilterSafeBrowsingClient::CheckResult::ToTracedValue() const {
   auto value = std::make_unique<base::trace_event::TracedValue>();
   value->SetInteger("request_id", request_id);
-  value->SetInteger("threat_type", threat_type);
+  value->SetInteger("threat_type", static_cast<int>(threat_type));
   value->SetValue("threat_metadata", threat_metadata.ToTracedValue().get());
   value->SetInteger("duration (us)",
                     (base::TimeTicks::Now() - start_time).InMicroseconds());
